Add Pokémon split names and types data for fusion generation
- Introduced POKEMON_SPLIT_NAMES to facilitate fusion name generation for Pokémon.
- Added POKEMON_NAMES to generate full Pokémon names from split names.
- Created POKEMON_TYPES object to map Pokémon IDs to their respective types.
- Data covers Pokémon from the first generation up to the fifth generation.
