+++ src/data/pokemon-names.js
@@ -1,0 +1,516 @@
+// Pokemon split names for fusion generation
+// Each Pokemon is represented as [prefix, suffix] to enable fusion name generation
+// Index corresponds to Pokemon ID (starting at 1, index 0 is empty placeholder)
+
+const POKEMON_SPLIT_NAMES = [
+  ['', ''],
+  ['Bulba', 'saur'],
+  ['Ivy', 'saur'],
+  ['Venu', 'saur'],
+  ['Char', 'mander'],
+  ['Char', 'meleon'],
+  ['Char', 'izard'],
+  ['Squirt', 'tle'],
+  ['War', 'tortle'],
+  ['Blast', 'toise'],
+  ['Cater', 'pie'],
+  ['Meta', 'pod'],
+  ['Butter', 'free'],
+  ['Wee', 'dle'],
+  ['Kak', 'una'],
+  ['Bee', 'drill'],
+  ['Pid', 'gey'],
+  ['Pidge', 'otto'],
+  ['Pid', 'geot'],
+  ['Rat', 'tata'],
+  ['Rat', 'icate'],
+  ['Spear', 'row'],
+  ['Fear', 'row'],
+  ['Ek', 'ans'],
+  ['Ar', 'bok'],
+  ['Pika', 'chu'],
+  ['Rai', 'chu'],
+  ['Sand', 'shrew'],
+  ['Sand', 'slash'],
+  ['Nido', 'ran'],
+  ['Nido', 'rina'],
+  ['Nido', 'queen'],
+  ['Nido', 'ran'],
+  ['Nido', 'rino'],
+  ['Nido', 'king'],
+  ['Clef', 'fairy'],
+  ['Clef', 'fable'],
+  ['Vul', 'pix'],
+  ['Nine', 'tales'],
+  ['Jiggly', 'puff'],
+  ['Wiggly', 'tuff'],
+  ['Zu', 'bat'],
+  ['Gol', 'bat'],
+  ['Odd', 'ish'],
+  ['Gloo', 'oom'],
+  ['Vile', 'plume'],
+  ['Pa', 'ras'],
+  ['Para', 'sect'],
+  ['Veno', 'nat'],
+  ['Veno', 'moth'],
+  ['Dig', 'lett'],
+  ['Dug', 'trio'],
+  ['Meow', 'th'],
+  ['Per', 'sian'],
+  ['Psy', 'duck'],
+  ['Gol', 'duck'],
+  ['Man', 'key'],
+  ['Prime', 'ape'],
+  ['Growl', 'lithe'],
+  ['Arca', 'nine'],
+  ['Poli', 'wag'],
+  ['Poli', 'whirl'],
+  ['Poli', 'wrath'],
+  ['Ab', 'ra'],
+  ['Kada', 'bra'],
+  ['Ala', 'kazam'],
+  ['Ma', 'chop'],
+  ['Ma', 'choke'],
+  ['Ma', 'champ'],
+  ['Bell', 'sprout'],
+  ['Weepin', 'bell'],
+  ['Victree', 'bell'],
+  ['Tenta', 'cool'],
+  ['Tenta', 'cruel'],
+  ['Geo', 'dude'],
+  ['Grave', 'ler'],
+  ['Go', 'lem'],
+  ['Pony', 'ta'],
+  ['Rapi', 'dash'],
+  ['Slow', 'poke'],
+  ['Slow', 'bro'],
+  ['Magne', 'mite'],
+  ['Magne', 'ton'],
+  ['Far', 'fetchd'],
+  ['Do', 'duo'],
+  ['Do', 'drio'],
+  ['See', 'eel'],
+  ['Dew', 'gong'],
+  ['Gri', 'mer'],
+  ['Mu', 'uk'],
+  ['Shell', 'der'],
+  ['Cloy', 'ster'],
+  ['Gas', 'tly'],
+  ['Haun', 'ter'],
+  ['Gen', 'gar'],
+  ['On', 'nix'],
+  ['Drow', 'zee'],
+  ['Hyp', 'no'],
+  ['Krab', 'by'],
+  ['King', 'ler'],
+  ['Volt', 'orb'],
+  ['Electr', 'ode'],
+  ['Exegg', 'cute'],
+  ['Exeggu', 'tor'],
+  ['Cu', 'bone'],
+  ['Maro', 'wak'],
+  ['Hitmon', 'lee'],
+  ['Hitmon', 'chan'],
+  ['Licki', 'tung'],
+  ['Koff', 'fing'],
+  ['Wee', 'zing'],
+  ['Rhy', 'horn'],
+  ['Rhy', 'don'],
+  ['Chan', 'sey'],
+  ['Tan', 'gela'],
+  ['Kangas', 'khan'],
+  ['Hor', 'sea'],
+  ['Sea', 'dra'],
+  ['Gol', 'deen'],
+  ['Sea', 'king'],
+  ['Star', 'yu'],
+  ['Star', 'mie'],
+  ['Mr. ', 'mime'],
+  ['Scy', 'ther'],
+  ['Jyn', 'nx'],
+  ['Electa', 'buzz'],
+  ['Mag', 'mar'],
+  ['Pin', 'sir'],
+  ['Tau', 'ros'],
+  ['Magi', 'karp'],
+  ['Gyara', 'dos'],
+  ['Lap', 'ras'],
+  ['Dit', 'to'],
+  ['Ee', 'vee'],
+  ['Vapor', 'eon'],
+  ['Jolt', 'eon'],
+  ['Fla', 'reon'],
+  ['Pory', 'gon'],
+  ['Oma', 'nyte'],
+  ['Oma', 'star'],
+  ['Kabu', 'to'],
+  ['Kabu', 'tops'],
+  ['Aero', 'dactyl'],
+  ['Snor', 'lax'],
+  ['Arti', 'cuno'],
+  ['Zap', 'dos'],
+  ['Mol', 'tres'],
+  ['Dra', 'tini'],
+  ['Drago', 'nair'],
+  ['Drago', 'nite'],
+  ['Mew', 'two'],
+  ['Mew', 'ew'],
+  ['Chiko', 'rita'],
+  ['Bay', 'leef'],
+  ['Mega', 'nium'],
+  ['Cynda', 'quil'],
+  ['Qui', 'lava'],
+  ['Typh', 'losion'],
+  ['Toto', 'dile'],
+  ['Croco', 'naw'],
+  ['Fera', 'ligatr'],
+  ['Sen', 'tret'],
+  ['Fur', 'ret'],
+  ['Hoot', 'hoot'],
+  ['Noct', 'owl'],
+  ['Ledy', 'ba'],
+  ['Led', 'ian'],
+  ['Spina', 'rak'],
+  ['Aria', 'dos'],
+  ['Cro', 'bat'],
+  ['Chin', 'chou'],
+  ['Lan', 'turn'],
+  ['Pi', 'chu'],
+  ['Clef', 'fa'],
+  ['Iggly', 'buff'],
+  ['Toge', 'pi'],
+  ['Toge', 'tic'],
+  ['Na', 'tu'],
+  ['Xa', 'tu'],
+  ['Ma', 'reep'],
+  ['Flaa', 'ffy'],
+  ['Ampha', 'ros'],
+  ['Bell', 'ossom'],
+  ['Ma', 'rill'],
+  ['Azuma', 'rill'],
+  ['Sudo', 'woodo'],
+  ['Poli', 'toed'],
+  ['Hop', 'pip'],
+  ['Skip', 'loom'],
+  ['Jump', 'luff'],
+  ['Ai', 'pom'],
+  ['Sun', 'kern'],
+  ['Sun', 'flora'],
+  ['Yan', 'ma'],
+  ['Woo', 'per'],
+  ['Quag', 'sire'],
+  ['Esp', 'eon'],
+  ['Umb', 'reon'],
+  ['Mur', 'krow'],
+  ['Slow', 'king'],
+  ['Mis', 'dreavus'],
+  ['Un', 'own'],
+  ['Wob', 'buffet'],
+  ['Gira', 'farig'],
+  ['Pine', 'co'],
+  ['Forre', 'tress'],
+  ['Dun', 'sparce'],
+  ['Gli', 'gar'],
+  ['Stee', 'lix'],
+  ['Snub', 'bull'],
+  ['Gran', 'bull'],
+  ['Qwil', 'fish'],
+  ['Sci', 'zor'],
+  ['Shu', 'ckle'],
+  ['Hera', 'cross'],
+  ['Snea', 'sel'],
+  ['Teddi', 'ursa'],
+  ['Ursa', 'ring'],
+  ['Slug', 'gma'],
+  ['Mag', 'cargo'],
+  ['Swi', 'nub'],
+  ['Pilo', 'swine'],
+  ['Cor', 'sola'],
+  ['Remo', 'raid'],
+  ['Octi', 'llery'],
+  ['Deli', 'bird'],
+  ['Man', 'tine'],
+  ['Skar', 'mory'],
+  ['Houn', 'dour'],
+  ['Houn', 'doom'],
+  ['King', 'dra'],
+  ['Phan', 'py'],
+  ['Don', 'phan'],
+  ['Pory', 'gon2'],
+  ['Stan', 'tler'],
+  ['Smear', 'gle'],
+  ['Ty', 'rogue'],
+  ['Hitmon', 'top'],
+  ['Smoo', 'chum'],
+  ['Ele', 'kid'],
+  ['Mag', 'by'],
+  ['Mil', 'tank'],
+  ['Blis', 'sey'],
+  ['Rai', 'kou'],
+  ['En', 'tei'],
+  ['Sui', 'cune'],
+  ['Larvi', 'tar'],
+  ['Pupi', 'tar'],
+  ['Tyran', 'itar'],
+  ['Lu', 'gia'],
+  ['Ho-', 'oh'],
+  ['Cele', 'bi'],
+  ['Azu', 'rill'],
+  ['Wy', 'naut'],
+  ['Ambi', 'pom'],
+  ['Mis', 'magius'],
+  ['Honch', 'krow'],
+  ['Bon', 'sly'],
+  ['Mime', ' Jr.'],
+  ['Happi', 'piny'],
+  ['Munch', 'lax'],
+  ['Man', 'tyke'],
+  ['Weav', 'vile'],
+  ['Magne', 'zone'],
+  ['Licki', 'licky'],
+  ['Rhy', 'perior'],
+  ['Tan', 'growth'],
+  ['Electi', 'vire'],
+  ['Magmor', 'mortar'],
+  ['Toge', 'kiss'],
+  ['Yan', 'mega'],
+  ['Leaf', 'feon'],
+  ['Glace', 'ceon'],
+  ['Glis', 'cor'],
+  ['Mamo', 'swine'],
+  ['Pory', 'gon-Z'],
+  ['Tree', 'cko'],
+  ['Gro', 'vyle'],
+  ['Scep', 'tile'],
+  ['Torch', 'chic'],
+  ['Com', 'busken'],
+  ['Bla', 'ziken'],
+  ['Mud', 'kip'],
+  ['Marsh', 'tomp'],
+  ['Swamp', 'pert'],
+  ['Ral', 'ts'],
+  ['Kir', 'lia'],
+  ['Garde', 'voir'],
+  ['Gall', 'lade'],
+  ['Shed', 'inja'],
+  ['Kec', 'eon'],
+  ['Bel', 'dum'],
+  ['Met', 'tang'],
+  ['Meta', 'gross'],
+  ['Bi', 'doof'],
+  ['Spiri', 'tomb'],
+  ['Luca', 'cario'],
+  ['Gib', 'ble'],
+  ['Gab', 'bite'],
+  ['Gar', 'chomp'],
+  ['Maw', 'wile'],
+  ['Lil', 'leep'],
+  ['Crad', 'dily'],
+  ['Ano', 'rith'],
+  ['Arm', 'aldo'],
+  ['Cran', 'idos'],
+  ['Ram', 'pardos'],
+  ['Shiel', 'don'],
+  ['Bastio', 'don'],
+  ['Sla', 'king'],
+  ['Ab', 'sol'],
+  ['Dus', 'skull'],
+  ['Dusc', 'clops'],
+  ['Dusk', 'noir'],
+  ['Wail', 'lord'],
+  ['Arc', 'ceus'],
+  ['Tur', 'wig'],
+  ['Gro', 'tle'],
+  ['Tor', 'terra'],
+  ['Chim', 'char'],
+  ['Mon', 'ferno'],
+  ['Infer', 'nape'],
+  ['Pip', 'lup'],
+  ['Prin', 'plup'],
+  ['Empo', 'leon'],
+  ['Nose', 'pass'],
+  ['Probo', 'pass'],
+  ['Hon', 'edge'],
+  ['Doub', 'blade'],
+  ['Aegi', 'slash'],
+  ['Pawn', 'iard'],
+  ['Bi', 'sharp'],
+  ['Lux', 'ray'],
+  ['Agg', 'gron'],
+  ['Fly', 'gon'],
+  ['Milo', 'tic'],
+  ['Sala', 'mence'],
+  ['Klink', 'klang'],
+  ['Zoro', 'ark'],
+  ['Syl', 'veon'],
+  ['Kyo', 'ogre'],
+  ['Grou', 'don'],
+  ['Ray', 'quaza'],
+  ['Dial', 'alga'],
+  ['Palk', 'kia'],
+  ['Gira', 'tina'],
+  ['Regi', 'gigas'],
+  ['Dark', 'rai'],
+  ['Gene', 'esect'],
+  ['Reshi', 'ram'],
+  ['Zek', 'rom'],
+  ['Kyu', 'rem'],
+  ['Rose', 'rade'],
+  ['Drifb', 'blim'],
+  ['Lop', 'punny'],
+  ['Bre', 'loom'],
+  ['Nin', 'jask'],
+  ['Ban', 'nette'],
+  ['Ro', 'tom'],
+  ['Reuni', 'niclus'],
+  ['Whimsi', 'cott'],
+  ['Krooko', 'dile'],
+  ['Cofa', 'grigus'],
+  ['Galvan', 'tula'],
+  ['Ferro', 'thorn'],
+  ['Lit', 'wick'],
+  ['Lamp', 'pent'],
+  ['Chandel', 'lure'],
+  ['Hax', 'xorus'],
+  ['Golu', 'lurk'],
+  ['Pyuku', 'muku'],
+  ['Klef', 'ki'],
+  ['Talon', 'flame'],
+  ['Mimi', 'kyu'],
+  ['Volca', 'rona'],
+  ['Dei', 'no'],
+  ['Zwei', 'lous'],
+  ['Hy', 'dreigon'],
+  ['La', 'tias'],
+  ['La', 'tios'],
+  ['Deo', 'xys'],
+  ['Ji', 'rachi'],
+  ['Nin', 'cada'],
+  ['Bi', 'barel'],
+  ['Rio', 'olu'],
+  ['Sla', 'koth'],
+  ['Vigo', 'roth'],
+  ['Wail', 'mer'],
+  ['Shi', 'nx'],
+  ['Lu', 'xio'],
+  ['Ar', 'on'],
+  ['Lair', 'ron'],
+  ['Trap', 'inch'],
+  ['Vibra', 'va'],
+  ['Fee', 'bas'],
+  ['Ba', 'agon'],
+  ['Shel', 'elgon'],
+  ['Kli', 'ink'],
+  ['Kla', 'ang'],
+  ['Zo', 'rua'],
+  ['Bu', 'dew'],
+  ['Rose', 'elia'],
+  ['Drifl', 'loon'],
+  ['Bun', 'eary'],
+  ['Shroom', 'ish'],
+  ['Shup', 'pet'],
+  ['Solo', 'sis'],
+  ['Duo', 'sion'],
+  ['Cotton', 'onee'],
+  ['Sandi', 'ile'],
+  ['Kroko', 'rok'],
+  ['Ya', 'mask'],
+  ['Jol', 'tik'],
+  ['Ferro', 'seed'],
+  ['Axe', 'xew'],
+  ['Frax', 'xure'],
+  ['Gole', 'olett'],
+  ['Fletch', 'ling'],
+  ['Fletch', 'inder'],
+  ['Larv', 'esta'],
+  ['Stun', 'fisk'],
+  ['Sabl', 'eye'],
+  ['Veni', 'pede'],
+  ['Whirli', 'pede'],
+  ['Scoli', 'pede'],
+  ['Ty', 'runt'],
+  ['Tyran', 'trum'],
+  ['Sno', 'runt'],
+  ['Gla', 'lie'],
+  ['Fros', 'lass'],
+  ['Ori', 'corio'],
+  ['Ori', 'corio'],
+  ['Ori', 'corio'],
+  ['Ori', 'corio'],
+  ['Trub', 'bish'],
+  ['Garb', 'odor'],
+  ['Car', 'vanha'],
+  ['Sharp', 'edo'],
+  ['Phant', 'tump'],
+  ['Tre', 'venant'],
+  ['Noi', 'bat'],
+  ['Noi', 'vern'],
+  ['Swab', 'lu'],
+  ['Alta', 'ria'],
+  ['Goo', 'my'],
+  ['Sli', 'goo'],
+  ['Goo', 'dra'],
+  ['Regi', 'rock'],
+  ['Regi', 'ice'],
+  ['Regi', 'steel'],
+  ['Necro', 'zma'],
+  ['Stuff', 'ful'],
+  ['Be', 'wear'],
+  ['Dhel', 'mise'],
+  ['Mar', 'eanie'],
+  ['Tox', 'apex'],
+  ['Hawl', 'lucha'],
+  ['Cac', 'nea'],
+  ['Cact', 'turne'],
+  ['Sandy', 'gast'],
+  ['Palo', 'sand'],
+  ['Ama', 'aura'],
+  ['Auro', 'rorus'],
+  ['Rock', 'ruff'],
+  ['Lycan', 'roc'],
+  ['Lycan', 'roc'],
+  ['Melo', 'etta'],
+  ['Melo', 'etta'],
+  ['Cress', 'selia'],
+  ['Brux', 'ish'],
+  ['Necro', 'zma'],
+  ['Jang', 'mo-o'],
+  ['Haka', 'mo-o'],
+  ['Kom', 'mo-o'],
+  ['Wim', 'pod'],
+  ['Goli', 'sopod'],
+  ['Fo', 'mantis'],
+  ['Lu', 'rantis'],
+  ['Car', 'bink'],
+  ['Ches', 'pin'],
+  ['Quil', 'ladin'],
+  ['Ches', 'naught'],
+  ['Fenne', 'kin'],
+  ['Brai', 'xen'],
+  ['Del', 'phox'],
+  ['Froa', 'kie'],
+  ['Frog', 'adier'],
+  ['Gre', 'ninja'],
+  ['Tor', 'koal'],
+  ['Pump', 'kaboo'],
+  ['Gour', 'geist'],
+  ['Swir', 'lix'],
+  ['Slur', 'puff'],
+  ['Scra', 'ggy'],
+  ['Scraf', 'ty'],
+  ['Lo', 'tad'],
+  ['Lom', 'bre'],
+  ['Ludi', 'colo'],
+  ['Mini', 'or'],
+  ['Mini', 'or'],
+  ['Dian', 'cie'],
+  ['Luv', 'disc'],
+];
+
+// Generate full Pokemon names list from split names for compatibility
+const POKEMON_NAMES = POKEMON_SPLIT_NAMES.slice(1).map(parts => parts.join(''));
+
+module.exports = {
+  POKEMON_SPLIT_NAMES,
+  POKEMON_NAMES,
+};
