+++ src/data/pokemon-types.js
@@ -1,0 +1,510 @@
+// Pokemon type data by ID for local type information
+// Key: Pokemon ID (1-501), Value: Array of types
+
+const POKEMON_TYPES = {
+  1: ['GRASS', 'POISON'],
+  2: ['GRASS', 'POISON'],
+  3: ['GRASS', 'POISON'],
+  4: ['FIRE'],
+  5: ['FIRE'],
+  6: ['FIRE', 'FLYING'],
+  7: ['WATER'],
+  8: ['WATER'],
+  9: ['WATER'],
+  10: ['BUG'],
+  11: ['BUG'],
+  12: ['BUG', 'FLYING'],
+  13: ['BUG', 'POISON'],
+  14: ['BUG', 'POISON'],
+  15: ['BUG', 'POISON'],
+  16: ['NORMAL', 'FLYING'],
+  17: ['NORMAL', 'FLYING'],
+  18: ['NORMAL', 'FLYING'],
+  19: ['NORMAL'],
+  20: ['NORMAL'],
+  21: ['NORMAL', 'FLYING'],
+  22: ['NORMAL', 'FLYING'],
+  23: ['POISON'],
+  24: ['POISON'],
+  25: ['ELECTRIC'],
+  26: ['ELECTRIC'],
+  27: ['GROUND'],
+  28: ['GROUND'],
+  29: ['POISON'],
+  30: ['POISON'],
+  31: ['POISON', 'GROUND'],
+  32: ['POISON'],
+  33: ['POISON'],
+  34: ['POISON', 'GROUND'],
+  35: ['FAIRY'],
+  36: ['FAIRY'],
+  37: ['FIRE'],
+  38: ['FIRE'],
+  39: ['NORMAL', 'FAIRY'],
+  40: ['NORMAL', 'FAIRY'],
+  41: ['POISON', 'FLYING'],
+  42: ['POISON', 'FLYING'],
+  43: ['GRASS', 'POISON'],
+  44: ['GRASS', 'POISON'],
+  45: ['GRASS', 'POISON'],
+  46: ['BUG', 'GRASS'],
+  47: ['BUG', 'GRASS'],
+  48: ['BUG', 'POISON'],
+  49: ['BUG', 'POISON'],
+  50: ['GROUND'],
+  51: ['GROUND'],
+  52: ['NORMAL'],
+  53: ['NORMAL'],
+  54: ['WATER'],
+  55: ['WATER'],
+  56: ['FIGHTING'],
+  57: ['FIGHTING'],
+  58: ['FIRE'],
+  59: ['FIRE'],
+  60: ['WATER'],
+  61: ['WATER'],
+  62: ['WATER', 'FIGHTING'],
+  63: ['PSYCHIC'],
+  64: ['PSYCHIC'],
+  65: ['PSYCHIC'],
+  66: ['FIGHTING'],
+  67: ['FIGHTING'],
+  68: ['FIGHTING'],
+  69: ['GRASS', 'POISON'],
+  70: ['GRASS', 'POISON'],
+  71: ['GRASS', 'POISON'],
+  72: ['WATER', 'POISON'],
+  73: ['WATER', 'POISON'],
+  74: ['ROCK', 'GROUND'],
+  75: ['ROCK', 'GROUND'],
+  76: ['ROCK', 'GROUND'],
+  77: ['FIRE'],
+  78: ['FIRE'],
+  79: ['WATER', 'PSYCHIC'],
+  80: ['WATER', 'PSYCHIC'],
+  81: ['STEEL', 'ELECTRIC'],
+  82: ['STEEL', 'ELECTRIC'],
+  83: ['NORMAL', 'FLYING'],
+  84: ['NORMAL', 'FLYING'],
+  85: ['NORMAL', 'FLYING'],
+  86: ['WATER'],
+  87: ['WATER', 'ICE'],
+  88: ['POISON'],
+  89: ['POISON'],
+  90: ['WATER'],
+  91: ['WATER', 'ICE'],
+  92: ['GHOST', 'POISON'],
+  93: ['GHOST', 'POISON'],
+  94: ['GHOST', 'POISON'],
+  95: ['ROCK', 'GROUND'],
+  96: ['PSYCHIC'],
+  97: ['PSYCHIC'],
+  98: ['WATER'],
+  99: ['WATER'],
+  100: ['ELECTRIC'],
+  101: ['ELECTRIC'],
+  102: ['GRASS', 'PSYCHIC'],
+  103: ['GRASS', 'PSYCHIC'],
+  104: ['GROUND'],
+  105: ['GROUND'],
+  106: ['FIGHTING'],
+  107: ['FIGHTING'],
+  108: ['NORMAL'],
+  109: ['POISON'],
+  110: ['POISON'],
+  111: ['GROUND', 'ROCK'],
+  112: ['GROUND', 'ROCK'],
+  113: ['NORMAL'],
+  114: ['GRASS'],
+  115: ['NORMAL'],
+  116: ['WATER'],
+  117: ['WATER'],
+  118: ['WATER'],
+  119: ['WATER'],
+  120: ['WATER'],
+  121: ['WATER', 'PSYCHIC'],
+  122: ['PSYCHIC', 'FAIRY'],
+  123: ['BUG', 'FLYING'],
+  124: ['ICE', 'PSYCHIC'],
+  125: ['ELECTRIC'],
+  126: ['FIRE'],
+  127: ['BUG'],
+  128: ['NORMAL'],
+  129: ['WATER'],
+  130: ['WATER', 'FLYING'],
+  131: ['WATER', 'ICE'],
+  132: ['NORMAL'],
+  133: ['NORMAL'],
+  134: ['WATER'],
+  135: ['ELECTRIC'],
+  136: ['FIRE'],
+  137: ['NORMAL'],
+  138: ['ROCK', 'WATER'],
+  139: ['ROCK', 'WATER'],
+  140: ['ROCK', 'WATER'],
+  141: ['ROCK', 'WATER'],
+  142: ['ROCK', 'FLYING'],
+  143: ['NORMAL'],
+  144: ['ICE', 'FLYING'],
+  145: ['ELECTRIC', 'FLYING'],
+  146: ['FIRE', 'FLYING'],
+  147: ['DRAGON'],
+  148: ['DRAGON'],
+  149: ['DRAGON', 'FLYING'],
+  150: ['PSYCHIC'],
+  151: ['PSYCHIC'],
+  152: ['GRASS'],
+  153: ['GRASS'],
+  154: ['GRASS'],
+  155: ['FIRE'],
+  156: ['FIRE'],
+  157: ['FIRE'],
+  158: ['WATER'],
+  159: ['WATER'],
+  160: ['WATER'],
+  161: ['NORMAL'],
+  162: ['NORMAL'],
+  163: ['NORMAL', 'FLYING'],
+  164: ['NORMAL', 'FLYING'],
+  165: ['BUG', 'FLYING'],
+  166: ['BUG', 'FLYING'],
+  167: ['BUG', 'POISON'],
+  168: ['BUG', 'POISON'],
+  169: ['POISON', 'FLYING'],
+  170: ['WATER', 'ELECTRIC'],
+  171: ['WATER', 'ELECTRIC'],
+  172: ['ELECTRIC'],
+  173: ['FAIRY'],
+  174: ['NORMAL', 'FAIRY'],
+  175: ['FAIRY'],
+  176: ['FAIRY', 'FLYING'],
+  177: ['PSYCHIC', 'FLYING'],
+  178: ['PSYCHIC', 'FLYING'],
+  179: ['ELECTRIC'],
+  180: ['ELECTRIC'],
+  181: ['ELECTRIC'],
+  182: ['GRASS'],
+  183: ['WATER', 'FAIRY'],
+  184: ['WATER', 'FAIRY'],
+  185: ['ROCK'],
+  186: ['WATER'],
+  187: ['GRASS', 'FLYING'],
+  188: ['GRASS', 'FLYING'],
+  189: ['GRASS', 'FLYING'],
+  190: ['NORMAL'],
+  191: ['GRASS'],
+  192: ['GRASS'],
+  193: ['BUG', 'FLYING'],
+  194: ['WATER', 'GROUND'],
+  195: ['WATER', 'GROUND'],
+  196: ['PSYCHIC'],
+  197: ['DARK'],
+  198: ['DARK', 'FLYING'],
+  199: ['WATER', 'PSYCHIC'],
+  200: ['GHOST'],
+  201: ['PSYCHIC'],
+  202: ['PSYCHIC'],
+  203: ['NORMAL', 'PSYCHIC'],
+  204: ['BUG'],
+  205: ['BUG', 'STEEL'],
+  206: ['NORMAL'],
+  207: ['GROUND', 'FLYING'],
+  208: ['STEEL', 'GROUND'],
+  209: ['FAIRY'],
+  210: ['FAIRY'],
+  211: ['WATER', 'POISON'],
+  212: ['BUG', 'STEEL'],
+  213: ['BUG', 'ROCK'],
+  214: ['BUG', 'FIGHTING'],
+  215: ['DARK', 'ICE'],
+  216: ['NORMAL'],
+  217: ['NORMAL'],
+  218: ['FIRE'],
+  219: ['FIRE', 'ROCK'],
+  220: ['ICE', 'GROUND'],
+  221: ['ICE', 'GROUND'],
+  222: ['WATER', 'ROCK'],
+  223: ['WATER'],
+  224: ['WATER'],
+  225: ['ICE', 'FLYING'],
+  226: ['WATER', 'FLYING'],
+  227: ['STEEL', 'FLYING'],
+  228: ['DARK', 'FIRE'],
+  229: ['DARK', 'FIRE'],
+  230: ['WATER', 'DRAGON'],
+  231: ['GROUND'],
+  232: ['GROUND'],
+  233: ['NORMAL'],
+  234: ['NORMAL'],
+  235: ['NORMAL'],
+  236: ['FIGHTING'],
+  237: ['FIGHTING'],
+  238: ['ICE', 'PSYCHIC'],
+  239: ['ELECTRIC'],
+  240: ['FIRE'],
+  241: ['NORMAL'],
+  242: ['NORMAL'],
+  243: ['ELECTRIC'],
+  244: ['FIRE'],
+  245: ['WATER'],
+  246: ['ROCK', 'GROUND'],
+  247: ['ROCK', 'GROUND'],
+  248: ['ROCK', 'DARK'],
+  249: ['PSYCHIC', 'FLYING'],
+  250: ['FIRE', 'FLYING'],
+  251: ['PSYCHIC', 'GRASS'],
+  252: ['NORMAL', 'FAIRY'],
+  253: ['PSYCHIC'],
+  254: ['NORMAL'],
+  255: ['GHOST'],
+  256: ['DARK', 'FLYING'],
+  257: ['ROCK'],
+  258: ['PSYCHIC', 'FAIRY'],
+  259: ['NORMAL'],
+  260: ['NORMAL'],
+  261: ['WATER', 'FLYING'],
+  262: ['DARK', 'ICE'],
+  263: ['STEEL', 'ELECTRIC'],
+  264: ['NORMAL'],
+  265: ['GROUND', 'ROCK'],
+  266: ['GRASS'],
+  267: ['ELECTRIC'],
+  268: ['FIRE'],
+  269: ['FAIRY', 'FLYING'],
+  270: ['BUG', 'FLYING'],
+  271: ['GRASS'],
+  272: ['ICE'],
+  273: ['GROUND', 'FLYING'],
+  274: ['ICE', 'GROUND'],
+  275: ['NORMAL'],
+  276: ['GRASS'],
+  277: ['GRASS'],
+  278: ['GRASS'],
+  279: ['FIRE'],
+  280: ['FIRE', 'FIGHTING'],
+  281: ['FIRE', 'FIGHTING'],
+  282: ['WATER'],
+  283: ['WATER', 'GROUND'],
+  284: ['WATER', 'GROUND'],
+  285: ['PSYCHIC', 'FAIRY'],
+  286: ['PSYCHIC', 'FAIRY'],
+  287: ['PSYCHIC', 'FAIRY'],
+  288: ['PSYCHIC', 'FIGHTING'],
+  289: ['BUG', 'GHOST'],
+  290: ['NORMAL'],
+  291: ['STEEL', 'PSYCHIC'],
+  292: ['STEEL', 'PSYCHIC'],
+  293: ['STEEL', 'PSYCHIC'],
+  294: ['NORMAL'],
+  295: ['DARK', 'GHOST'],
+  296: ['FIGHTING', 'STEEL'],
+  297: ['DRAGON', 'GROUND'],
+  298: ['DRAGON', 'GROUND'],
+  299: ['DRAGON', 'GROUND'],
+  300: ['STEEL', 'FAIRY'],
+  301: ['ROCK', 'GRASS'],
+  302: ['ROCK', 'GRASS'],
+  303: ['ROCK', 'BUG'],
+  304: ['ROCK', 'BUG'],
+  305: ['ROCK'],
+  306: ['ROCK'],
+  307: ['ROCK', 'STEEL'],
+  308: ['ROCK', 'STEEL'],
+  309: ['NORMAL'],
+  310: ['DARK'],
+  311: ['GHOST'],
+  312: ['GHOST'],
+  313: ['GHOST'],
+  314: ['WATER'],
+  315: ['NORMAL'],
+  316: ['GRASS'],
+  317: ['GRASS'],
+  318: ['GRASS', 'GROUND'],
+  319: ['FIRE'],
+  320: ['FIRE', 'FIGHTING'],
+  321: ['FIRE', 'FIGHTING'],
+  322: ['WATER'],
+  323: ['WATER'],
+  324: ['WATER', 'STEEL'],
+  325: ['ROCK'],
+  326: ['ROCK', 'STEEL'],
+  327: ['STEEL', 'GHOST'],
+  328: ['STEEL', 'GHOST'],
+  329: ['STEEL', 'GHOST'],
+  330: ['DARK', 'STEEL'],
+  331: ['DARK', 'STEEL'],
+  332: ['ELECTRIC'],
+  333: ['STEEL', 'ROCK'],
+  334: ['GROUND', 'DRAGON'],
+  335: ['WATER'],
+  336: ['DRAGON', 'FLYING'],
+  337: ['STEEL'],
+  338: ['DARK'],
+  339: ['FAIRY'],
+  340: ['WATER'],
+  341: ['GROUND'],
+  342: ['DRAGON', 'FLYING'],
+  343: ['STEEL', 'DRAGON'],
+  344: ['WATER', 'DRAGON'],
+  345: ['GHOST', 'DRAGON'],
+  346: ['NORMAL'],
+  347: ['DARK'],
+  348: ['BUG', 'STEEL'],
+  349: ['DRAGON', 'FIRE'],
+  350: ['DRAGON', 'ELECTRIC'],
+  351: ['DRAGON', 'ICE'],
+  352: ['GRASS', 'POISON'],
+  353: ['GHOST', 'FLYING'],
+  354: ['NORMAL'],
+  355: ['GRASS', 'FIGHTING'],
+  356: ['BUG', 'FLYING'],
+  357: ['GHOST'],
+  358: ['ELECTRIC', 'GHOST'],
+  359: ['PSYCHIC'],
+  360: ['GRASS', 'FAIRY'],
+  361: ['GROUND', 'DARK'],
+  362: ['GHOST'],
+  363: ['BUG', 'ELECTRIC'],
+  364: ['STEEL', 'GRASS'],
+  365: ['GHOST', 'FIRE'],
+  366: ['GHOST', 'FIRE'],
+  367: ['GHOST', 'FIRE'],
+  368: ['DRAGON'],
+  369: ['GROUND', 'GHOST'],
+  370: ['WATER'],
+  371: ['STEEL', 'FAIRY'],
+  372: ['FIRE', 'FLYING'],
+  373: ['GHOST', 'FAIRY'],
+  374: ['BUG', 'FIRE'],
+  375: ['DARK', 'DRAGON'],
+  376: ['DARK', 'DRAGON'],
+  377: ['DARK', 'DRAGON'],
+  378: ['DRAGON', 'PSYCHIC'],
+  379: ['DRAGON', 'PSYCHIC'],
+  380: ['PSYCHIC'],
+  381: ['STEEL', 'PSYCHIC'],
+  382: ['BUG', 'GROUND'],
+  383: ['NORMAL', 'WATER'],
+  384: ['FIGHTING'],
+  385: ['NORMAL'],
+  386: ['NORMAL'],
+  387: ['WATER'],
+  388: ['ELECTRIC'],
+  389: ['ELECTRIC'],
+  390: ['STEEL', 'ROCK'],
+  391: ['STEEL', 'ROCK'],
+  392: ['GROUND'],
+  393: ['GROUND', 'DRAGON'],
+  394: ['WATER'],
+  395: ['DRAGON'],
+  396: ['DRAGON'],
+  397: ['STEEL'],
+  398: ['STEEL'],
+  399: ['DARK'],
+  400: ['GRASS', 'POISON'],
+  401: ['GRASS', 'POISON'],
+  402: ['GHOST', 'FLYING'],
+  403: ['NORMAL'],
+  404: ['GRASS'],
+  405: ['GHOST'],
+  406: ['PSYCHIC'],
+  407: ['PSYCHIC'],
+  408: ['GRASS', 'FAIRY'],
+  409: ['GROUND', 'DARK'],
+  410: ['GROUND', 'DARK'],
+  411: ['GHOST'],
+  412: ['BUG', 'ELECTRIC'],
+  413: ['STEEL', 'GRASS'],
+  414: ['DRAGON'],
+  415: ['DRAGON'],
+  416: ['GROUND', 'GHOST'],
+  417: ['NORMAL', 'FLYING'],
+  418: ['FIRE', 'FLYING'],
+  419: ['BUG', 'FIRE'],
+  420: ['GROUND', 'ELECTRIC'],
+  421: ['DARK', 'GHOST'],
+  422: ['BUG', 'POISON'],
+  423: ['BUG', 'POISON'],
+  424: ['BUG', 'POISON'],
+  425: ['ROCK', 'DRAGON'],
+  426: ['ROCK', 'DRAGON'],
+  427: ['ICE'],
+  428: ['ICE'],
+  429: ['ICE', 'GHOST'],
+  430: ['FIRE', 'FLYING'],
+  431: ['POISON'],
+  432: ['POISON'],
+  433: ['WATER', 'DARK'],
+  434: ['WATER', 'DARK'],
+  435: ['GRASS', 'GHOST'],
+  436: ['GRASS', 'GHOST'],
+  437: ['FLYING', 'DRAGON'],
+  438: ['FLYING', 'DRAGON'],
+  439: ['NORMAL', 'FLYING'],
+  440: ['DRAGON', 'FLYING'],
+  441: ['DRAGON'],
+  442: ['DRAGON'],
+  443: ['DRAGON'],
+  444: ['ROCK'],
+  445: ['ICE'],
+  446: ['STEEL'],
+  447: ['PSYCHIC'],
+  448: ['NORMAL', 'FIGHTING'],
+  449: ['NORMAL', 'FIGHTING'],
+  450: ['GHOST', 'GRASS'],
+  451: ['POISON', 'WATER'],
+  452: ['POISON', 'WATER'],
+  453: ['FIGHTING', 'FLYING'],
+  454: ['GRASS'],
+  455: ['GRASS', 'DARK'],
+  456: ['GROUND', 'GHOST'],
+  457: ['GROUND', 'GHOST'],
+  458: ['ROCK', 'ICE'],
+  459: ['ROCK', 'ICE'],
+  460: ['ROCK'],
+  461: ['ROCK'],
+  462: ['ROCK'],
+  463: ['NORMAL', 'PSYCHIC'],
+  464: ['NORMAL', 'FIGHTING'],
+  465: ['PSYCHIC'],
+  466: ['WATER', 'PSYCHIC'],
+  467: ['PSYCHIC', 'DRAGON'],
+  468: ['DRAGON'],
+  469: ['DRAGON', 'FIGHTING'],
+  470: ['DRAGON', 'FIGHTING'],
+  471: ['BUG', 'WATER'],
+  472: ['BUG', 'WATER'],
+  473: ['GRASS'],
+  474: ['GRASS'],
+  475: ['ROCK', 'FAIRY'],
+  476: ['GRASS'],
+  477: ['GRASS'],
+  478: ['GRASS', 'FIGHTING'],
+  479: ['FIRE'],
+  480: ['FIRE'],
+  481: ['FIRE', 'PSYCHIC'],
+  482: ['WATER'],
+  483: ['WATER'],
+  484: ['WATER', 'DARK'],
+  485: ['FIRE'],
+  486: ['GHOST', 'GRASS'],
+  487: ['GHOST', 'GRASS'],
+  488: ['FAIRY'],
+  489: ['FAIRY'],
+  490: ['DARK', 'FIGHTING'],
+  491: ['DARK', 'FIGHTING'],
+  492: ['WATER', 'GRASS'],
+  493: ['WATER', 'GRASS'],
+  494: ['WATER', 'GRASS'],
+  495: ['ROCK', 'FLYING'],
+  496: ['ROCK', 'FLYING'],
+  497: ['ROCK', 'FAIRY'],
+  498: ['WATER'],
+  499: ['WATER'],
+  500: ['WATER'],
+  501: ['WATER'],
+};
+
+module.exports = {
+  POKEMON_TYPES,
+};
